refactor: move isChromiumWindow to utils
